Fix for netcombo.com.br
Whitelist entry for http://www.netcombo.com.br/  (the website of the major Brazilian ISP). The current filter `||navdmp.com^` makes it impossible to get past the form on the frontpage. Type in the name of a city, e.g. Brasília, and press either of the buttons. Nothing happens because of that filter.