Whitelisted trackjs.com
removed block of `capture.trackjs.com` and whitelisted the subdomains of `trackjs.com`. 

TrackJS is an error monitoring service that records errors from user sessions and sends the data back to the developers of the web site. We do not aggregate data, sell data, or provide advertisements. We do not track users between customers either.